# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Open Observability Agent API
  version: 0.1.0
  description: Multi-tenant observability and evidence-backed agent operations API.
servers:
  - url: http://localhost:8080
security:
  - ApiKey: []
paths:
  /health:
    get:
      security: []
      responses:
        '200': { description: OK }
  /v1/observations:
    post:
      summary: Ingest an observation
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/Observation' } } }
      responses:
        '201': { description: Created }
        '400': { description: Invalid observation }
    get:
      summary: Query tenant observations
      parameters:
        - { name: tenant_id, in: query, required: true, schema: { type: string, format: uuid } }
        - { name: name, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string, enum: [ok, error] } }
        - { name: page, in: query, schema: { type: integer, minimum: 1 } }
        - { name: page_size, in: query, schema: { type: integer, minimum: 1, maximum: 1000 } }
      responses: { '200': { description: Observation page } }
  /v1/diagnostics:
    post:
      summary: Generate findings from tenant observations
      responses: { '200': { description: Findings } }
  /v1/observations/batch:
    post:
      summary: Ingest a validated batch of observations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [observations]
              properties:
                observations: { type: array, minItems: 1, maxItems: 1000, items: { $ref: '#/components/schemas/Observation' } }
      responses:
        '201': { description: Number of observations accepted }
        '400': { description: Batch rejected before append }
  /v1/agent/plan:
    post:
      summary: Plan an evidence-backed agent action
      responses: { '200': { description: Agent decision } }
  /v1/agent/execute:
    post:
      summary: Execute an allowlisted tool with approval state
      responses:
        '200': { description: Tool execution result }
        '401': { description: Missing API key }
  /v1/model/complete:
    post:
      summary: Complete a model request and emit model observation plus token usage
      responses:
        '200': { description: Model response and observation }
  /v1/usage:
    post:
      summary: Record billable usage
      responses: { '201': { description: Recorded } }
    get:
      summary: Read usage for a tenant and period
      responses: { '200': { description: Usage snapshot } }
  /v1/billing/quote:
    post:
      summary: Calculate a monthly quote in integer cents
      responses: { '200': { description: Billing quote } }
components:
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    TenantId: { type: object, required: [0] }
    Observation:
      type: object
      required: [id, tenant_id, trace_id, span_id, kind, name, status, started_at_ms, duration_ms, attributes]
      properties:
        id: { type: string, format: uuid }
        tenant_id: { type: object }
        trace_id: { type: string }
        span_id: { type: string }
        kind: { type: string, enum: [Agent, Tool, Model, Workflow, Http] }
        name: { type: string, minLength: 1 }
        status: { type: string, enum: [Ok, Error] }
        started_at_ms: { type: integer, format: int64 }
        duration_ms: { type: integer, format: int64, minimum: 0 }
        attributes: { type: object, additionalProperties: { type: string } }